Fixed issue #14
Tweepy 1.9 introduced a buffer parameter to make the stream less CPU intensive.
This has the side effect that users with less busy twitter feeds had a less
than realtime stream. I set the BUFFER_SIZE back to the default of 1. I still
want to make it either configurable from weechat or add a function that
autodetects a sane buffer size.