First version of realtime search.
I've implemented the first realtime search version, it's far from stable but it
works to a point, when trying some of the very busy searches it seems to crash
but that's a problem I will have to solve tomorrow. Will merge with master once
it feels stable and ready for my main client.